crank/requirements.txt
solana-py==0.34.0
solders==0.21.0
anchorpy==0.20.1
httpx==0.27.0
apscheduler==3.10.4
python-dotenv==1.0.1
pandas==2.2.2
